script/requirements.txt
psycopg2
requests
python-dotenv